Simple refactors
