+++ packages/api/src/ontology/index.ts
@@ -42,8 +42,6 @@ export type {
   ValidPropertyTypes,
 } from "./Definition";
 export type { OntologyMetadata } from "./OntologyMetadata";
-export type { OsdkObject } from "./OsdkObject";
-export type { OsdkObjectFrom } from "./OsdkObjectFrom";
 export type {
   AggregationKeyDataType,
   ObjectQueryDataType,
