+++ packages/legacy-client/src/client/objectSets/createObjectSetTerminalLoadStep.ts
@@ -23,7 +23,7 @@ import type { ObjectSetDefinition } from "../baseTypes";
 import type { ObjectSetTerminalLoadStep } from "../interfaces";
 import { loadAllObjects } from "../net/loadObjects";
 import { loadObjectsPage } from "../net/loadObjectsPage";
-import type { OsdkLegacyObjectFrom } from "../OsdkObject";
+import type { OsdkLegacyObjectFrom } from "../OsdkLegacyObject";
 import type { OrderByClause } from "./filters";
 
 export function createObjectSetTerminalLoadStep<
