+++ packages/legacy-client/src/client/objectSets/mapPropertiesToGroupByProperties.ts
@@ -23,7 +23,7 @@ import {
   StringGroupBy,
   TimestampGroupBy,
 } from "../objectSets/aggregations";
-import type { OsdkLegacyObjectFrom } from "../OsdkObject";
+import type { OsdkLegacyObjectFrom } from "../OsdkLegacyObject";
 import { isReservedKeyword } from "../utils/reservedKeywords";
 
 export function mapPropertiesToGroupByProperties<
