+++ packages/legacy-client/src/client/objectSets/mapPropertiesToOrderBy.ts
@@ -17,7 +17,7 @@
 import type { ObjectTypesFrom, OntologyDefinition } from "@osdk/api";
 import type { ObjectSetOrderByArg } from "../interfaces/ordering";
 import { OrderByOption } from "../objectSets/filters";
-import type { OsdkLegacyObjectFrom } from "../OsdkObject";
+import type { OsdkLegacyObjectFrom } from "../OsdkLegacyObject";
 import { isReservedKeyword } from "../utils/reservedKeywords";
 
 export function mapPropertiesToOrderBy<
